+++ fhir_auth/example/android/app/src/main/kotlin/com/example/example/MainActivity.kt
@@ -1,4 +1,4 @@
-package dev.fhirfli.authdemo
+package dev.fhirfli.fhirauth
 
 import io.flutter.embedding.android.FlutterActivity